+++ example/lib/demos/export.dart
@@ -1,0 +1,3 @@
+export 'rtoast&rSnackbar_demo.dart';
+export 'jtextplus_demo.dart';
+export 'jmarquee_demo.dart';