+++ lib/apis/src/datetime_api/timeago.dart
@@ -24,7 +24,7 @@ class Timeago {
     this.showWeeks = false,
     this.showMonths = true,
     this.showYears = true,
-  }) : this.duration = endAt.difference(startAt);
+  }) : duration = endAt.difference(startAt);
 
   Timeago.since(DateTime since,
       {this.code,
@@ -35,7 +35,7 @@ class Timeago {
       this.showWeeks = false,
       this.showMonths = true,
       this.showYears = true})
-      : this.duration = DateTime.now().difference(since);
+      : duration = DateTime.now().difference(since);
 
   Timeago.fromDuration(this.duration,
       {this.code,
@@ -54,27 +54,28 @@ class Timeago {
       String Function(DateTimeFormat)? yearFormat}) {
     var dtl = DateTimeLang.dateTimeLang(code);
 
-    if (showYears && duration.inYears > 0)
+    if (showYears && duration.inYears > 0) {
       return (yearFormat == null)
-          ? '${duration.inYears} ${isFull ? dtl.YEARS : dtl.SHORTYEARS} ${dtl.TIMEAGESUFFIX}'
+          ? '${duration.inYears} ${isFull ? dtl.years : dtl.shortyears} ${dtl.timeagosuffix}'
           : yearFormat(DateTimeFormat(DateTime.now().subtract(duration),
               seprator: seprator, code: code));
-    else if (showMonths && duration.inMonths > 0)
+    } else if (showMonths && duration.inMonths > 0) {
       return (monthFormat == null)
-          ? '${duration.inMonths} ${isFull ? dtl.MONTHS.NAME : dtl.SHORTMONTHS.NAME} ${dtl.TIMEAGESUFFIX}'
+          ? '${duration.inMonths} ${isFull ? dtl.months.NAME : dtl.shortmonths.NAME} ${dtl.timeagosuffix}'
           : monthFormat(DateTimeFormat(DateTime.now().subtract(duration),
               seprator: seprator, code: code));
-    else if (showWeeks && duration.inWeeks > 0)
-      return '${duration.inWeeks} ${isFull ? dtl.WEEKDAYS.NAME : dtl.SHORTWEEKDAYS.NAME} ${dtl.TIMEAGESUFFIX}';
-    else if (showDays && duration.inDays > 0)
-      return '${duration.inDays} ${isFull ? dtl.DAYS : dtl.SHORTDAYS} ${dtl.TIMEAGESUFFIX}';
-    else if (showHours && duration.inHours > 0)
-      return '${duration.inHours} ${isFull ? dtl.HOURS : dtl.SHORTHOURS} ${dtl.TIMEAGESUFFIX}';
-    else if (showMinutes && duration.inMinutes > 0)
-      return '${duration.inMinutes} ${isFull ? dtl.MINUTES : dtl.SHORTMINUTES} ${dtl.TIMEAGESUFFIX}';
-    else if (showSeconds && duration.inSeconds > 10)
-      return '${duration.inSeconds} ${isFull ? dtl.SECONDS : dtl.SHORTSECONDS} ${dtl.TIMEAGESUFFIX}';
-    else
-      return dtl.JUSTNOW;
+    } else if (showWeeks && duration.inWeeks > 0) {
+      return '${duration.inWeeks} ${isFull ? dtl.weekdays.NAME : dtl.shortweekdays.NAME} ${dtl.timeagosuffix}';
+    } else if (showDays && duration.inDays > 0) {
+      return '${duration.inDays} ${isFull ? dtl.days : dtl.shortdays} ${dtl.timeagosuffix}';
+    } else if (showHours && duration.inHours > 0) {
+      return '${duration.inHours} ${isFull ? dtl.hours : dtl.shorthours} ${dtl.timeagosuffix}';
+    } else if (showMinutes && duration.inMinutes > 0) {
+      return '${duration.inMinutes} ${isFull ? dtl.minutes : dtl.shortminutes} ${dtl.timeagosuffix}';
+    } else if (showSeconds && duration.inSeconds > 10) {
+      return '${duration.inSeconds} ${isFull ? dtl.seconds : dtl.shortseconds} ${dtl.timeagosuffix}';
+    } else {
+      return dtl.justnow;
+    }
   }
 }