+++ lib/widgets/src/jtextplus/export_jtextpluse.dart
@@ -1,0 +1,2 @@
+export 'jtextplus.dart';
+export 'models/jtextstyle.dart';