+++ CHANGELOG.md
@@ -1,3 +1,31 @@
+## 0.0.5
+
+🎉 Upgrade to `jars ^0.0.5`
+
+- ✅ Fix documentation
+
+## 0.0.4
+
+🎉 Upgrade to `jars ^0.0.4`
+
+- ✅ Extensions Upgrade
+- ✅ Upgrade documentation
+
+## 0.0.3
+
+🎉 Upgrade to `jars ^0.0.3`
+
+- ✅ JTextPlus Widget Added
+- ✅ JMarquee Widget Added
+- ✅ Upgrade DateTime API
+- ✅ Upgrade documentation
+
+## 0.0.2
+
+🎉 Upgrade to `jars ^0.0.2`
+
+- ✅ Fix documentation
+
 ## 0.0.1
 
-* TODO: Describe initial release.
+🎉 initial release.